fixed  bottom  sheet  player
